+++ EleFits/EleFits/impl/BintableColumns.hpp
@@ -241,23 +241,6 @@ void BintableColumns::insert_null(Linx::Index index, const TInfo& info) const
   // TODO to Cfitsio
 }
 
-template <typename TInfo>
-void BintableColumns::init(const TInfo& info, Linx::Index index) const
-{
-  m_edit();
-  auto name = Fits::String::to_char_ptr(info.name);
-  auto tform = Fits::String::to_char_ptr(Cfitsio::TypeCode<typename TInfo::Value>::tform(info.repeat_count()));
-  int status = 0;
-  int cfitsio_index = index == -1 ? Cfitsio::BintableIo::column_count(m_fptr) + 1 : index + 1;
-  fits_insert_col(m_fptr, cfitsio_index, name.get(), tform.get(), &status);
-  Cfitsio::CfitsioError::may_throw(status, m_fptr, "Cannot init new column: #" + std::to_string(index));
-  if (info.unit != "") {
-    const Record<std::string> record {"TUNIT" + std::to_string(cfitsio_index), info.unit, "", "physical unit of field"};
-    Cfitsio::HeaderIo::update_record(m_fptr, record);
-  }
-  // TODO to Cfitsio
-}
-
 // write_segment
 
 template <typename TColumn>